:zap: Redefine viring for validate-on-input validators
Previously, the virgin definition is that no input is in pending state. It is better to say that no input is in changed state while pending. This way, if the user makes changes and undoes those changes within the same input session, the input is taken to not have been messed with. This is consistent behavior as validate-on-confirm if the user makes changes and undoes those changes, thereby confirming the exact same value.