:bug: Fix validation issues when combining on-input vs on-confirm validators
Problem 1: Previously, on-input validators would not repeat themselves on confirm and the on-confirm validators will not wait on them, so if on-confirm validators are only enabled pending the on-input validator result, the on-confirm validators will not always validate correctly. The fix is to make sure that on-input validator is waited on when validating in confirm mode.

Problem 2: Previously, we ignore older validation if the confirm ID changes. But when we validate on-input, we should ignore when we have a new validation cycle. In fact, this should generally be happening. So we uncouple the confirmID (which is used to determine whether we should still auto-advance the form) from the validation ID (which is used to determine whether the validation cycle should continue).